Move decorator from main.py to decorators.py
